// quizonline-frontend/src/app/api/generated/model/course-invite-status-enum.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `pending` - Pending * `accepted` - Accepted * `declined` - Declined * `revoked` - Revoked * `expired` - Expired
 */
export enum CourseInviteStatusEnumDto {

    Pending = 'pending',

    Accepted = 'accepted',

    Declined = 'declined',

    Revoked = 'revoked',

    Expired = 'expired'
}